requirements.txt
fastapi
uvicorn
msal
httpx
python-dotenv
pydantic
sqlalchemy
jwt
